requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
numpy==1.26.3
faker==22.2.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-cov==4.1.0
flake8==7.0.0
